Move createItemToDropFromEntity WorldUtils classto
More logical location for createItemToDropFromEntity to exist at.
